Projects/Large_image_model/requirements.txt
ipykernel
google-generativeai
python-dotenv
ipywidgets
pandas
numpy
seaborn
matplotlib
openai
openpyxl
streamlit